requirements.txt
streamlit>=1.20.0
python-dotenv>=1.0.0
requests>=2.25.0
# Optional / conditional packages:
openai>=0.27.0
azure-storage-blob>=12.14.1
